feat: Create Notification within each once reservation
This commit has a lot of things, First I've added a new token holder to
use it when A user is needed and shouldn't be retrieved from frontend.
Every time a Once reservation is created a notification is saved.
Every time a Reservation Change of state, a new notification is also
saved.

There are a few things I should improved, but I think that this base should
be good enough at the moment.